+++ src/tab-size/tab-size.js
@@ -1,3 +1,5 @@
+import addStyleToPage from '../add-style'
+
 export default function setTabSize(numSpaces) {
     const cssRule = createCssRule(numSpaces)
     addStyleToPage(cssRule)
@@ -10,12 +12,4 @@ function createCssRule(numSpaces) {
         }
     `
     return cssRule
-}
-
-// inject CSS <style> into <head> of page
-function addStyleToPage(cssRule) {
-    const css = document.createElement('style')
-    css.type = 'text/css'
-    css.appendChild(document.createTextNode(cssRule))
-    document.getElementsByTagName('head')[0].appendChild(css)
 }